Refactor: EnrichPerformance, StatementData 패키지 이동
